# season-1/2.vertical-vs-horizontal-scaling/contracts/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Scaling Lab Baseline API",
    "version": "0.1.0",
    "description": "Source-grounded compatibility baseline at 8cad37f. Known defects are recorded, not endorsed. Full canonical contract is Express; comparison implementations declare capabilities separately."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5102"
    }
  ],
  "paths": {
...
    "/cart": {
      "get": {
        "operationId": "listCart",
        "summary": "listCart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Baseline success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartList"
                }
              }
            }
          },
          "400": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "description": "Baseline: invalid or nonpositive values default to 20; values over 100 clamp."
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ]
      },
      "post": {
        "operationId": "addCart",
        "summary": "addCart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Baseline success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "201": {
            "description": "Baseline success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "400": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartInput"
              }
            }
          }
        }
      }
    },
    "/cart/{id}": {
      "get": {
        "operationId": "getCart",
        "summary": "getCart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Baseline success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "400": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ]
      },
      "put": {
        "operationId": "updateCart",
        "summary": "updateCart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Baseline success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "400": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartPatch"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteCart",
        "summary": "deleteCart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Baseline success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedCart"
                }
              }
            }
          },
          "400": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ]
      }
    },
    "/cart/{id}/quantity": {
      "patch": {
        "operationId": "adjustCart",
        "summary": "adjustCart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Baseline success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartAdjustment"
                }
              }
            }
          },
          "400": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityDelta"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "description",
          "price",
          "stock",
          "category",
          "imageUrl"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "stock": {
            "type": "integer",
            "minimum": 0
          },
          "category": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Cart": {
        "type": "object",
        "required": [
          "_id",
          "userId",
          "productId",
          "qty"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$"
          },
          "userId": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$"
          },
          "productId": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-fA-F0-9]{24}$",
                "nullable": true
              },
              {
                "$ref": "#/components/schemas/Product"
              }
            ]
          },
          "qty": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "CartInput": {
        "type": "object",
        "required": [
          "productId",
          "qty"
        ],
        "properties": {
          "productId": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$"
          },
          "qty": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "CartPatch": {
        "type": "object",
        "required": [
          "qty"
        ],
        "properties": {
          "qty": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "QuantityDelta": {
        "type": "object",
        "required": [
          "delta"
        ],
        "properties": {
          "delta": {
            "type": "integer",
            "not": {
              "enum": [
                0
              ]
            }
          }
        }
      },
      "CartList": {
        "type": "object",
        "required": [
          "items",
          "nextCursor",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cart"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "total": {
            "type": "number"
          }
        }
      },
...
      "DeletedCart": {
        "type": "object",
        "required": [
          "deleted"
        ],
        "properties": {
          "deleted": {
            "$ref": "#/components/schemas/Cart"
          }
        }
      },
...
      "CartAdjustment": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/Cart"
          },
          {
            "$ref": "#/components/schemas/DeletedCart"
          }
        ]
      },
...
    }
  }
}